+++ .gitignore
@@ -22,11 +22,13 @@ DerivedData
 *.xcuserstate
 project.xcworkspace
 
-# Android/IJ
+# Android/IntelliJ
 #
+build/
 .idea
 .gradle
 local.properties
+*.iml
 
 # Web
 #
@@ -36,7 +38,20 @@ local.properties
 #
 node_modules/
 npm-debug.log
-src/testKey.js
+yarn-error.log
+
+# BUCK
+buck-out/
+\.buckd/
 *.keystore
-main.jsbundle
-ios/noder/build
+
+# fastlane
+#
+# It is recommended to not store the screenshots in the git repo. Instead, use fastlane to re-generate the
+# screenshots whenever they are needed.
+# For more information about the recommended setup visit:
+# https://github.com/fastlane/fastlane/blob/master/fastlane/docs/Gitignore.md
+
+fastlane/report.xml
+fastlane/Preview.html
+fastlane/screenshots
