+++ android/app/src/main/java/com/noder/MainActivity.java
@@ -1,15 +1,6 @@
 package com.lingyong.me.noder;
 
 import com.facebook.react.ReactActivity;
-import com.microsoft.codepush.react.CodePush;
-import com.oblador.vectoricons.VectorIconsPackage;
-import com.lwansbrough.RCTCamera.RCTCameraPackage;
-import com.eguma.barcodescanner.BarcodeScannerPackage;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-
-import java.util.Arrays;
-import java.util.List;
 
 public class MainActivity extends ReactActivity {
 
@@ -20,38 +11,5 @@ public class MainActivity extends ReactActivity {
     @Override
     protected String getMainComponentName() {
         return "noder";
-    }
-
-    /**
-     * Returns whether dev mode should be enabled.
-     * This enables e.g. the dev menu.
-     */
-    @Override
-    protected boolean getUseDeveloperSupport() {
-        return BuildConfig.DEBUG;
-    }
-
-    // 2. Override the getJSBundleFile method in order to let
-    // the CodePush runtime determine where to get the JS
-    // bundle location from on each app start
-    @Override
-    protected String getJSBundleFile() {
-            return CodePush.getBundleUrl();
-    }
-
-
-    /**
-     * A list of packages used by the app. If the app uses additional views
-     * or modules besides the default ones, add more packages here.
-     */
-    @Override
-    protected List<ReactPackage> getPackages() {
-        return Arrays.<ReactPackage>asList(
-            new MainReactPackage(),
-            new CodePush("Q2A8khx6JV4mXXcS0usR0LipDz0Y410YoZxlZ", this, BuildConfig.DEBUG),
-            new VectorIconsPackage(),
-            new RCTCameraPackage(),
-            new BarcodeScannerPackage()
-        );
     }
 }
