+++ app.json
@@ -1,0 +1,4 @@
+{
+  "name": "noder",
+  "displayName": "noder"
+}
