+++ package.json
@@ -1,81 +1,89 @@
 {
-  "name": "noder",
-  "version": "1.0.5",
-  "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "android": "node node_modules/react-native/local-cli/cli.js run-android",
-    "ios": "node node_modules/react-native/local-cli/cli.js run-ios",
-    "web": "node node_modules/react-web/local-cli/cli.js start web/webpack.config.js",
-    "log": "adb logcat *:S ReactNative:V ReactNativeJS:V",
-    "build-web": "node node_modules/react-web/local-cli/cli.js bundle web/webpack.config.js",
-    "build-ios": "react-native unbundle --entry-file index.ios.js --platform ios --dev false",
-    "build-android": "cd android && ./gradlew assembleRelease && open app/build/outputs/apk && cd ..",
-    "checkversion": "node node_modules/fbjs-scripts/node/check-dev-engines.js package.json",
-    "postinstall": "npm run checkversion",
-    "push-android": "code-push release-react Noder android --deploymentName Staging",
-    "push-ios": "code-push release-react Noder ios --deploymentName Staging",
-    "push-android-prod": "code-push release-react Noder android --deploymentName Production",
-    "push-ios-prod": "code-push release-react Noder ios --deploymentName Production",
-    "push-key": "code-push deployment ls Noder -k"
-  },
-  "dependencies": {
-    "fetch-detector": "^1.0.0",
-    "fetch-ie8": "^1.4.3",
-    "flux-standard-action": "^0.6.1",
-    "lodash": "^4.13.1",
-    "markdown": "0.5.0",
-    "moment": "^2.13.0",
-    "query-string": "^4.2.2",
-    "react": "15.1.0",
-    "react-addons-pure-render-mixin": "15.1.0",
-    "react-dom": "15.1.0",
-    "react-native": "^0.28.0",
-    "react-native-barcodescanner": "^3.0.0",
-    "react-native-blur": "^1.0.0",
-    "react-native-button": "^1.6.0",
-    "react-native-camera": "git+https://github.com/lwansbrough/react-native-camera.git",
-    "react-native-cli": "^1.0.0",
-    "react-native-code-push": "^1.12.2-beta",
-    "react-native-html-render": "^1.0.5",
-    "react-native-scrollable-tab-view": "^0.5.1",
-    "react-native-vector-icons": "^2.0.3",
-    "react-redux": "^4.4.5",
-    "react-web": "git+https://github.com/flyskywhy/react-web.git",
-    "redux": "^3.5.2",
-    "redux-actions": "^0.10.0",
-    "redux-logger": "^2.6.1",
-    "redux-promise": "^0.5.3",
-    "redux-thunk": "^2.1.0"
-  },
-  "devDependencies": {
-    "babel-eslint": "^6.1.0",
-    "babel-loader": "^6.2.4",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-1": "^6.5.0",
-    "coffee-script": "^1.9.2",
-    "dev-ip": "^1.0.1",
-    "eslint": "^2.13.1",
-    "eslint-plugin-react": "^6.2.0",
-    "file-loader": "^0.9.0",
-    "gulp": "^3.9.1",
-    "gulp-replace": "^0.5.4",
-    "gulp-util": "^3.0.4",
-    "haste-resolver-webpack-plugin": "^0.2.1",
-    "json-loader": "^0.5.4",
-    "react-hot-loader": "^1.3.0",
-    "react-native-cli": "^0.2.0",
-    "redux-devtools": "^3.3.1",
-    "run-sequence": "^1.2.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-html-plugin": "^0.1.1"
-  },
-  "devEngines": {
-    "node": ">= 4.x",
-    "npm": ">= 3.x"
-  },
-  "bundleId": "org.reactjs.native.example.noder"
+	"name": "noder",
+	"version": "1.0.5",
+	"scripts": {
+		"start": "node node_modules/react-native/local-cli/cli.js start",
+		"test": "jest",
+		"android": "node node_modules/react-native/local-cli/cli.js run-android",
+		"ios": "node node_modules/react-native/local-cli/cli.js run-ios",
+		"web": "node node_modules/react-web/local-cli/cli.js start web/webpack.config.js",
+		"log": "adb logcat *:S ReactNative:V ReactNativeJS:V",
+		"build-web": "node node_modules/react-web/local-cli/cli.js bundle web/webpack.config.js",
+		"build-ios": "react-native unbundle --entry-file index.ios.js --platform ios --dev false",
+		"build-android": "cd android && ./gradlew assembleRelease && open app/build/outputs/apk && cd ..",
+		"checkversion": "node node_modules/fbjs-scripts/node/check-dev-engines.js package.json",
+		"postinstall": "npm run checkversion",
+		"push-android": "code-push release-react Noder android --deploymentName Staging",
+		"push-ios": "code-push release-react Noder ios --deploymentName Staging",
+		"push-android-prod": "code-push release-react Noder android --deploymentName Production",
+		"push-ios-prod": "code-push release-react Noder ios --deploymentName Production",
+		"push-key": "code-push deployment ls Noder -k"
+	},
+	"dependencies": {
+		"fetch-detector": "^1.0.0",
+		"fetch-ie8": "^1.4.3",
+		"flux-standard-action": "^0.6.1",
+		"lodash": "^4.13.1",
+		"markdown": "0.5.0",
+		"moment": "^2.13.0",
+		"query-string": "^4.2.2",
+		"react": "16.0.0-alpha.6",
+		"react-addons-pure-render-mixin": "15.6.2",
+		"react-dom": "16.0.0-alpha.2",
+		"react-native": "0.43.4",
+		"react-native-barcodescanner": "^3.0.0",
+		"react-native-blur": "^1.0.0",
+		"react-native-button": "^1.6.0",
+		"react-native-camera": "git+https://github.com/lwansbrough/react-native-camera.git",
+		"react-native-cli": "^1.0.0",
+		"react-native-code-push": "^1.12.2-beta",
+		"react-native-html-render": "^1.0.5",
+		"react-native-scrollable-tab-view": "^0.5.1",
+		"react-native-vector-icons": "^2.0.3",
+		"react-redux": "^4.4.5",
+		"react-web": "git+https://github.com/flyskywhy/react-web.git",
+		"redux": "^3.5.2",
+		"redux-actions": "^0.10.0",
+		"redux-logger": "^2.6.1",
+		"redux-promise": "^0.5.3",
+		"redux-thunk": "^2.1.0"
+	},
+	"devDependencies": {
+		"babel-eslint": "^6.1.0",
+		"babel-jest": "21.2.0",
+		"babel-loader": "^6.2.4",
+		"babel-polyfill": "^6.13.0",
+		"babel-preset-es2015": "^6.6.0",
+		"babel-preset-react": "^6.5.0",
+		"babel-preset-react-native": "4.0.0",
+		"babel-preset-stage-1": "^6.5.0",
+		"coffee-script": "^1.9.2",
+		"dev-ip": "^1.0.1",
+		"eslint": "^2.13.1",
+		"eslint-plugin-react": "^6.2.0",
+		"file-loader": "^0.9.0",
+		"gulp": "^3.9.1",
+		"gulp-replace": "^0.5.4",
+		"gulp-util": "^3.0.4",
+		"haste-resolver-webpack-plugin": "^0.2.1",
+		"jest": "21.2.1",
+		"json-loader": "^0.5.4",
+		"react-hot-loader": "^1.3.0",
+		"react-native-cli": "^0.2.0",
+		"react-test-renderer": "16.0.0-alpha.6",
+		"redux-devtools": "^3.3.1",
+		"run-sequence": "^1.2.1",
+		"url-loader": "^0.5.7",
+		"webpack": "^1.13.1",
+		"webpack-dev-server": "^1.14.1",
+		"webpack-html-plugin": "^0.1.1"
+	},
+	"jest": {
+		"preset": "react-native"
+	},
+	"devEngines": {
+		"node": ">= 4.x",
+		"npm": ">= 3.x"
+	},
+	"bundleId": "org.reactjs.native.example.noder"
 }
